+++ src/app/Components/Centerfunction.jsx
@@ -1,0 +1,39 @@
+import React, { Component } from "react";
+import Slider from "react-slick";
+
+function CenterMode() {
+    const settings = {
+        className: "center",
+        centerMode: true,
+        infinite: true,
+        centerPadding: "60px",
+        slidesToShow: 3,
+        speed: 500
+    };
+    return (
+        <div className="max-w-xl mx-auto">
+            <Slider {...settings}>
+                <div>
+                    <h3>1</h3>
+                </div>
+                <div>
+                    <h3>2</h3>
+                </div>
+                <div>
+                    <h3>3</h3>
+                </div>
+                <div>
+                    <h3>4</h3>
+                </div>
+                <div>
+                    <h3>5</h3>
+                </div>
+                <div>
+                    <h3>6</h3>
+                </div>
+            </Slider>
+        </div>
+    );
+}
+
+export default CenterMode;
